requirements.txt
streamlit
requests
pandas
python-dateutil